rename filenames EuclideanVector -> Vector
